pyproject.toml
[tool.poetry]
name = "langgraph-agents"
version = "0.1.0"
description = ""
authors = ["Connor (Cub) <connor@cubdigital.co.nz>"]
readme = "README.md"
packages = [{ include = "agents" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langgraph = "^0.2.35"
langchain-openai = "^0.2.2"
langchain = "^0.3.3"
langchain-core = "^0.3.10"
python-dotenv = "^1.0.1"
rich = "^13.9.2"
langchain-community = "^0.3.2"
duckduckgo-search = "^6.3.0"
arxiv = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
text_agent_test = "agents.text_agent:test"
travel_agent_test = "agents.travel_agent:test"
tts_agent_test = "agents.text_to_speech_agent:test"
research_agent_test = "agents.research_agent:test"
